fix SP payload of MIKEY messages (#827)
fill PayloadSPPolicyParamTypeAuthTagLen and
PayloadSPPolicyParamTypeSessionAuthKeyLen properly.
